CX-137: Only try to copy user and dev docs if folders exist

+++ install/cx/build/cxCustusXBuilder.py
@@ -128,16 +128,18 @@ class CustusXBuilder:
     def _moveUserDocsToStandardLocation(self):
         installer_path=self._getInitialInstallerPackagePath()
         source = '%s/%s' % (installer_path, "doc/html_pure")
-        dest = '%s/%s' % (self._getStandardInstallerPackagePath(), os.path.basename(source))
-        PrintFormatter.printInfo('Copying user doc files from [%s] to [%s]'%(source,dest))
-        shutil.copytree(source, dest)
+        if os.path.exists(source):
+            dest = '%s/%s' % (self._getStandardInstallerPackagePath(), os.path.basename(source))
+            PrintFormatter.printInfo('Copying user doc files from [%s] to [%s]'%(source,dest))
+            shutil.copytree(source, dest)
 
     def _moveDevDocsToStandardLocation(self):
         installer_path=self._getInitialInstallerPackagePath()
         source = '%s/%s' % (installer_path, "doc/html_dev")
-        dest = '%s/%s' % (self._getStandardInstallerPackagePath(), os.path.basename(source))
-        PrintFormatter.printInfo('Copying dev doc files from [%s] to [%s]'%(source,dest))
-        shutil.copytree(source, dest)
+        if os.path.exists(source):
+            dest = '%s/%s' % (self._getStandardInstallerPackagePath(), os.path.basename(source))
+            PrintFormatter.printInfo('Copying dev doc files from [%s] to [%s]'%(source,dest))
+            shutil.copytree(source, dest)
 
     def createInstallerObject(self, installer_path=None):    
         if installer_path==None:
